Make `spinalcord` a default model to solve `PyTorchStreamReader` error in `sct_run_batch` (#5076)
This is a simple little PR that will avoid PyTorchStreamReader errors
when running the sct_run_batch section of the SCT Course, since we'll no
longer attempt to download the model in parallel.

We could set other models as default, but I'm not sure it's worth the
additional bandwidth/storage usage on installation, when otherwise
they'll just be downloaded when running inference.

There's maybe an argument to pre-downloading any models that are
included in the course (to make sure all of our users have the necessary
model files *before* starting the course, to avoid delays during the
course)? Happy to discuss.

## Linked issues

Fixes #5036.