Clarify `ValueError` if the `-centerline` does not cover all the slices in the input mask (#5077)
## Description

This is just a tiny follow-up PR on
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/5075, which
clarifies `ValueError` if the `-centerline` does not cover all the
slices in the input mask. This can happen, for example, if the cord
segmentation is shorter than the canal segmentation.

Otherwise, I got an error mentioning the previous arg
`-angle-corr-centerline` (renamed to `-centerline` in
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/5075):

```
  File "code/spinalcordtoolbox/spinalcordtoolbox/process_seg.py", line 91, in compute_shape
    raise ValueError(
ValueError: The provided angle correction centerline does not cover slice(s) 1 of the input mask. Please supply a more extensive '-angle-corr-centerline', or disable angle correction ('-angle-corr 0').
```

I also searched for other `-angle-corr-centerline` occurrences and
didn't find any other except for in `CHANGES.md`.

## Linked issues

Follow-up:
https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/5075