Solve projection anomaly on anisotropic data (#5067)
## Description

This PR fixes an issue where label projection was incorrect on
anisotropic data. The problem occurred because the projection was
performed in pixel coordinates instead of real-world space. The greater
the anisotropy, the worse the projection became. The fix ensures that
projections now correctly account for the image resolution, making them
accurate regardless of anisotropy.

![ezgif-1f2a8071ab8f8b](https://github.com/user-attachments/assets/bba260d3-68d3-4246-a969-e73b5837cc7f)

## Linked issues

Fixes https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/5063
Fixes #5004

---------

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>