Update how `-discfile` intermediate images are saved (output directory, relative paths) (#5074)
## Description (fix 1: relative paths)

`Image.save` will fail if the output path contains non-existent parent
directories. This can occur for `-discfile` because we switch the base
directory to a temporary directory using `os.path.join` (which preserves
all of the parent directories).

This PR fixes the issue by using the `TempFolder.copy_from` method,
which ensures that all `-discfile` tempfile processing is done relative
to the tempdir (without any parent directories).

#### Discussion

- SCT has a few different functions/classes/methods involving tempdirs.
I'm not sure which method we currently recommend, but maybe we should
try to analyze + cull methods that lead to this kind of error, that way
new contributions can avoid this error in the future. I went with
`copy_from` since it feels correct to copy the inputs to a tempdir prior
to processing, and it lets us use `add_suffix` much more cleanly. But,
I'm open to other approaches as well.
- I'm not sure how to detect this kind of error elsewhere. It
specifically requires a relative path to trigger, but all of our tests
are written with absolute paths in mind (thanks to the `sct_test_data()`
function + the test data being in a different location than the working
directory).

## Description (fix 2: output directory)

Right now, these temporary files are being saved to the same directory
as the output `-csv` file. But, this isn't ideal, since we might want
the .csv to be in a different directory than the per-subject directories
(see:
https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/4070#issuecomment-3473957864).

So, this PR changes the output directory for these temporary files to be
the same as the input image.

## Linked issues

Fix 1: Fixes #5070.
Fix 2: Fixes a suggestion made in #5067.
Fix 2: Related to #4070.