`sct_deepseg.py`: Replace `NotImplementedError` w/ task-specific args
This prevents `-test-time-aug` from even being shown for `ivadomed` models, which I think is a better
approach.

I'm not sure if there's a way we can check if a model is ivadomed/nnunetv2 _before_ it's downloaded currently
(our model type checker only checks models that have already been downloaded), so I needed to implement
a new 'framework' metadata field for the models).