`test_cli.py`: Add new test that explicitly checks import of modules
This new test gets to the thrust of why we even had an "import time" test in the first place -- to catch
the import of expensive modules on new PRs. I figure that going straight to the source will help contributors
understand and fix their contributions while not running into any flakiness whatsoever.

Also, one benefit of taking this approach is that it prevents the "accumulation of small slowdowns" -- e.g.
if an import takes 300ms, it might not trip the old test by itself, but it would when combined with other
semi-slow imports. So, we catch them early before they accumulate and get merged.

I tried a few methods to figure out which modules are getting imported (e.g. the built-in modulefinder), but
they seemed to be broken not only due to upstream bugs (https://bugs.python.org/issue40350) but also due to
pytest causing some sort of infinite recursion. I tried both `modulefinder` and the spin-off `modulegraph`
but neither seemed to be able to figure out the imports.

The `sys.modules`-based approach seems to work well enough, even if `del` feels a bit awkward to attain
test isolation.